Update S1 tests
