+++ SentinelOneDeepVisibility/main.py
@@ -1,4 +1,4 @@
-from deep_visibility.base import SentinelOneDeepVisibilityModule
+from deep_visibility import SentinelOneDeepVisibilityModule
 from deep_visibility.connector_s3_logs import DeepVisibilityConnector
 
 if __name__ == "__main__":
